--- Assets/Scripts/SS3D/Animations/AirlockStateMachine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5a0a09942afdc84b8eafc1329b6b62b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
