--- Assets/Scripts/SS3D/Animations/SS3D.Animation.asmdef
@@ -1,14 +1,0 @@
-{
-    "name": "SS3D.Animation",
-    "rootNamespace": "",
-    "references": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
