--- Assets/Scripts/SS3D/Animations/SS3D.Animation.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7af3b582dde2ce94b87a2ae9e6f7aa1d
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
