+++ Assets/Scripts/SS3D/Systems/Furniture/AirlockStateMachine.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 27730ec8f794e9148be5975dc87faaef
+guid: e7efcf3f2a046b54c97cf93cbdf4704e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
