+++ Assets/Scripts/SS3D/Systems/SS3D.Systems.asmdef
@@ -15,14 +15,12 @@
         "GUID:d2bbca7254de5e746857b04cfb560876",
         "GUID:b763ff73a93a62c4ea128e32ed2cd0d6",
         "GUID:1476e4af5b8d89d4c9963efe314428b9",
-        "GUID:e66dbdfabf9e6a14ead1b347276e9bb7",
         "GUID:9e24947de15b9834991c9d8411ea37cf",
         "GUID:75469ad4d38634e559750d17036d5f7c",
         "GUID:eb124556ba20d60469be9771eb1f0360",
         "GUID:46e20dca70571ce4e817a6b5874168a0",
         "GUID:8979f103c1228714eb0115e440b2b33a",
-        "GUID:47588e3d37f573d458b0eaa6cc8109c2",
-        "GUID:7af3b582dde2ce94b87a2ae9e6f7aa1d"
+        "GUID:47588e3d37f573d458b0eaa6cc8109c2"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
